Activate XML Editor's toolbar when an object is selected
When introducing deferred_on_tree_select_row, not all paths on XmlTree::on_tree_select_row where covered. This lead to missing proper activation of the row in the XmlTree. The failure was that then the buttons sensitity was not set correctly and they were not activated.

Changed XmlTree::on_tree_select_row to now cover both options correctly:
- path if selection changes first on the desktop side (introduced with this commit)
- path if selection changes first on the Xml Editor side

Resolves: #2941